feat: apply linter
